# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "alex",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ACTIVE",
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "SUSPENDED"
              ]
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "example": "IMAM",
              "type": "string",
              "enum": [
                "SUPER_ADMIN",
                "MASJID_ADMIN",
                "IMAM",
                "COMMITTEE_MEMBER",
                "MEMBER"
              ]
            }
          },
          {
            "name": "masjidId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PaginatedResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Requires permission: platform.users.read",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "timestamp": "2026-04-01T00:00:00.000Z",
                    "path": "/api/v1/admin/users"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get users with pagination and optional search",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "example": {
                            "id": "usr_01HXYZ123",
                            "fullName": "Alex Johnson",
                            "status": "ACTIVE"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Requires permission: platform.users.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/status": {
      "patch": {
        "description": "INACTIVE and SUSPENDED also sign the user out of every session. SUPER_ADMIN accounts: 403 SUPER_ADMIN_IMMUTABLE. Unknown id: 404 USER_NOT_FOUND.",
        "operationId": "AdminController_updateUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "example": {
                            "id": "usr_01HXYZ123",
                            "status": "SUSPENDED"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Requires permission: platform.users.manage"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user status",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/roles": {
      "post": {
        "description": "Returns the same shape as GET /admin/users/:id. Only IMAM, COMMITTEE_MEMBER and MEMBER can be assigned (403 ROLE_NOT_ASSIGNABLE).",
        "operationId": "AdminController_assignRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Roles assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "example": {
                            "id": "usr_01HXYZ123",
                            "roles": [
                              "MASJID_ADMIN",
                              "IMAM"
                            ]
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Requires permission: platform.roles.assign"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the roles of a user",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/dashboard/summary": {
      "get": {
        "operationId": "AdminDashboardController_getSummary",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Requires permission: platform.users.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get optimized super admin dashboard summary",
        "tags": [
          "Admin Dashboard"
        ]
      }
    },
    "/admin/masjids": {
      "get": {
        "operationId": "AdminMasjidsController_listMasjids",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Jama Masjid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "SUSPENDED"
              ]
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Uttar Pradesh",
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "example": "India",
              "type": "string"
            }
          },
          {
            "name": "district",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Bareilly",
              "type": "string"
            }
          },
          {
            "name": "locality",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Barota",
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Requires permission: platform.masjids.manage"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get masjids with pagination and filters",
        "tags": [
          "Admin Masjids"
        ]
      }
    },
    "/admin/masjids/{id}": {
      "get": {
        "operationId": "AdminMasjidsController_getMasjid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Requires permission: platform.masjids.manage"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get masjid details by id",
        "tags": [
          "Admin Masjids"
        ]
      }
    },
    "/admin/masjids/{id}/status": {
      "patch": {
        "description": "REJECTED and SUSPENDED require a reason (400 VALIDATION_ERROR). Unknown id: 404 MASJID_NOT_FOUND.",
        "operationId": "AdminMasjidsController_updateMasjidStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMasjidStatusDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Requires permission: platform.masjids.manage"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update masjid status",
        "tags": [
          "Admin Masjids"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masjid Core API",
    "description": "Backend for the masjid management system. Every masjid route needs a permission; see src/access/permissions.ts.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from /auth/login/verify-otp"
      }
    },
    "schemas": {
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Request successful"
          },
          "data": {
            "type": "object",
            "description": "Payload returned by the endpoint",
            "example": {
              "id": "usr_123",
              "fullName": "Alex Johnson"
            }
          }
        },
        "required": [
          "success",
          "message",
          "data"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalItems": {
            "type": "number",
            "example": 43
          },
          "totalPages": {
            "type": "number",
            "example": 5
          },
          "hasPreviousPage": {
            "type": "boolean",
            "example": false
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "page",
          "limit",
          "totalItems",
          "totalPages",
          "hasPreviousPage",
          "hasNextPage"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "Page items. Item shape depends on the endpoint.",
            "example": []
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ],
            "example": "ACTIVE"
          }
        },
        "required": [
          "status"
        ]
      },
      "AssignUserRolesDto": {
        "type": "object",
        "properties": {
          "roleNames": {
            "example": [
              "MASJID_ADMIN",
              "IMAM"
            ],
            "description": "Role names to assign to the user",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roleNames"
        ]
      },
      "UpdateMasjidStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "SUSPENDED"
            ],
            "example": "APPROVED"
          },
          "reason": {
            "type": "string",
            "example": "Duplicate registration"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}